+++ hcs-sxc-java-examples/hcs-sxc-java-settlement-demo/src/main/java/com/hedera/hcsapp/CreateTopic.java
@@ -30,7 +30,7 @@ import lombok.extern.log4j.Log4j2;
 public class CreateTopic {
 
     public static void main(String[] args) throws Exception {
-        HCSCore hcsCore = HCSCore.INSTANCE.singletonInstanceDefault(0L);
+        HCSCore hcsCore = HCSCore.INSTANCE.singletonInstance();
         // create topics on HCS
         ConsensusTopicId topicId = new CreateHCSTopic(hcsCore)
                 .execute();