+++ hcs-sxc-java-examples/hcs-sxc-java-settlement-demo/src/main/java/com/hedera/hcsapp/Statics.java
@@ -1,0 +1,13 @@
+package com.hedera.hcsapp;
+
+
+public final class Statics {
+    private static AppData appData;
+
+    public static AppData getAppData() throws Exception {
+        if (appData == null) {
+            appData = new AppData();
+        }
+        return appData;
+    }
+}