+++ hcs-sxc-java-examples/hcs-sxc-java-settlement-demo/src/main/java/com/hedera/hcsapp/restclasses/CreditRest.java
@@ -24,9 +24,7 @@ import com.hedera.hcsapp.AppData;
 import com.hedera.hcsapp.States;
 import com.hedera.hcsapp.entities.Credit;
 import lombok.Data;
-import lombok.extern.log4j.Log4j2;
 
-@Log4j2
 @Data
 public final class CreditRest {
 
@@ -44,21 +42,17 @@ public final class CreditRest {
     private String topicId;
 
     public CreditRest(Credit credit, AppData appData) {
-        try {
-            this.threadId = credit.getThreadId();
-            this.applicationMessageId = credit.getApplicationMessageId();
-            this.payerName = credit.getPayerName();
-            this.recipientName = credit.getRecipientName();
-            this.reference = credit.getReference();
-            this.amount = credit.getAmount();
-            this.currency = credit.getCurrency();
-            this.additionalNotes = credit.getAdditionalNotes();
-            this.status = credit.getStatus();
-            this.createdDateTime = credit.getCreatedDate() + " " + credit.getCreatedTime();
-            this.topicId = appData.getHCSCore().getTopics().get(appData.getTopicIndex()).getTopic();
-            this.displayStatus = States.valueOf(this.status).getDisplayForCredit();
-        } catch (Exception e) {
-            log.error(e);
-        }
+        this.threadId = credit.getThreadId();
+        this.applicationMessageId = credit.getApplicationMessageId();
+        this.payerName = credit.getPayerName();
+        this.recipientName = credit.getRecipientName();
+        this.reference = credit.getReference();
+        this.amount = credit.getAmount();
+        this.currency = credit.getCurrency();
+        this.additionalNotes = credit.getAdditionalNotes();
+        this.status = credit.getStatus();
+        this.createdDateTime = credit.getCreatedDate() + " " + credit.getCreatedTime();
+        this.topicId = appData.getHCSCore().getTopics().get(appData.getTopicIndex()).getTopic();
+        this.displayStatus = States.valueOf(this.status).getDisplayForCredit();
     }
 }