+++ hcs-sxc-java-examples/hcs-sxc-java-simple-message-demo/src/main/java/com/hedera/hcsapp/CreateTopic.java
@@ -1,5 +1,7 @@
 package com.hedera.hcsapp;
 
+import com.hedera.hashgraph.sdk.account.AccountId;
+
 /*-
  * ‌
  * hcs-sxc-java
@@ -21,16 +23,25 @@ package com.hedera.hcsapp;
  */
 
 import com.hedera.hashgraph.sdk.consensus.ConsensusTopicId;
+import com.hedera.hashgraph.sdk.crypto.ed25519.Ed25519PrivateKey;
 import com.hedera.hcs.sxc.HCSCore;
 import com.hedera.hcs.sxc.consensus.CreateHCSTopic;
 
+import io.github.cdimascio.dotenv.Dotenv;
 import lombok.extern.log4j.Log4j2;
 
 @Log4j2
 public class CreateTopic {
 
     public static void main(String[] args) throws Exception {
-        HCSCore hcsCore = HCSCore.INSTANCE.singletonInstanceDefault(0);
+        Dotenv dotEnv = Dotenv.load();
+        Ed25519PrivateKey operatorKey = Ed25519PrivateKey.fromString(dotEnv.get("OPERATOR_KEY"));
+        AccountId operatorId = AccountId.fromString(dotEnv.get("OPERATOR_ID"));
+
+        HCSCore hcsCore = HCSCore.INSTANCE.singletonInstance("0")
+                .withOperatorAccountId(operatorId)
+                .withOperatorKey(operatorKey);
+        
         // create topics on HCS
         CreateHCSTopic createHCSTopic = new CreateHCSTopic(hcsCore);
         ConsensusTopicId topicId = createHCSTopic.execute();